from shreyaslad fork: fix: change vector representation from []float64 -> []float32

+++ request/pointsSearch.go
@@ -12,7 +12,7 @@ import (
 type PointsSearch struct {
 	CollectionName string              `json:"-"`
 	Consistency    *string             `json:"-"`
-	Vector         []float64           `json:"vector"`
+	Vector         []float32           `json:"vector"`
 	Filter         Filter              `json:"filter,omitempty"`
 	Params         *PointsSearchParams `json:"params,omitempty"`
 	Limit          int                 `json:"limit"`
